fix: :bug: Somes fix from review
